上传功能已完成了个

+++ src/main/java/com/kexin/admin/service/impl/MachineModelServiceImpl.java
@@ -4,12 +4,10 @@ import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.kexin.admin.entity.tables.DataupLog;
 import com.kexin.admin.entity.tables.MachineModel;
+import com.kexin.admin.entity.tables.Operation;
 import com.kexin.admin.entity.tables.Operator;
 import com.kexin.admin.entity.vo.Ftp;
-import com.kexin.admin.mapper.DataupLogMapper;
-import com.kexin.admin.mapper.LoginUserMapper;
-import com.kexin.admin.mapper.MachineModelMapper;
-import com.kexin.admin.mapper.OperatorMapper;
+import com.kexin.admin.mapper.*;
 import com.kexin.admin.service.MachineModelService;
 import com.kexin.common.util.FileUtil.FileUtil;
 import com.kexin.common.util.ResponseEty;
@@ -48,6 +46,15 @@ public class MachineModelServiceImpl extends ServiceImpl<MachineModelMapper, Mac
     @Resource
     OperatorMapper operatorMapper;//人员mapper
 
+    @Resource
+    MachineMapper machineMapper;//设备mapper
+
+    @Resource
+    OperationMapper operationMapper;//工序mapper
+
+    @Resource
+    ProductsMapper productsMapper;//产品mapper
+
     @Override
     public Integer machineModelCountByCode(String machineModelCode) {
         QueryWrapper<MachineModel> wrapper = new QueryWrapper<>();
@@ -77,12 +84,22 @@ public class MachineModelServiceImpl extends ServiceImpl<MachineModelMapper, Mac
     @Transactional(rollbackFor = Exception.class)
     public void updateMachineModel(MachineModel machineModel) {
 //        dropUserRolesByUserId(user.getLoginId());
+        if (machineModel.getUseFlag()){//启用
+            machineModel.setUseFlag(true);
+            machineModel.setEndDate(null);
+            machineModel.setStartDate(new Date());
+
+        }else{//禁用
+            machineModel.setUseFlag(false);
+            machineModel.setEndDate(new Date());
+        }
         baseMapper.updateById(machineModel);
     }
 
     @Override
     @Transactional(rollbackFor = Exception.class)
     public void deleteMachineModel(MachineModel machineModel) {
+
         baseMapper.deleteById(machineModel.getMachineModelId());
     }
 
@@ -119,6 +136,10 @@ public class MachineModelServiceImpl extends ServiceImpl<MachineModelMapper, Mac
             return ResponseEty.failure("只能上传一个文件");
         }
         MachineModel machineModel=baseMapper.selectById(addId);
+        machineModel.setMachine(machineMapper.selectById(machineModel.getMachineId()));
+        machineModel.setOperation(operationMapper.selectById(machineModel.getOperationId()));
+        machineModel.setProduct(productsMapper.selectById(machineModel.getProductId()));
+
         machineModel.setMachineModelNum(machineModel.getMachineModelNum()+1);//版本号+1
         String fileName=machineModel.getMachineModelNum()+rfilename;
         machineModel.setFileName(fileName);
@@ -135,6 +156,8 @@ public class MachineModelServiceImpl extends ServiceImpl<MachineModelMapper, Mac
 
         String s; // 本地文件的全地址
         String suffix;//文件的后缀
+        Boolean isSuccess = true;
+        String message="";
         for (int j=0;j<list.size();j++){
             FTPClient ftpClient = FTPUtil.connectFtpServer(ftp.getIpAddr(), ftp.getPort(), ftp.getUserName(), ftp.getPwd(), ftp.getEncoding());
 //            FTPUtil ftpUtil=new FTPUtil();
@@ -146,25 +169,96 @@ public class MachineModelServiceImpl extends ServiceImpl<MachineModelMapper, Mac
 
             Map map = FTPUtil.uploadMachineModelFiles(ftpClient, new File(s),ftp,machineModel);
 
-            Boolean isSuccess= (Boolean) map.get("success");
-            String message= (String) map.get("message");
+            isSuccess= (Boolean) map.get("success");
+            message= (String) map.get("message");
             if (isSuccess){//上传成功
                 FTPUtil.closeFTPConnect(ftpClient);
-                rfilename=ftp.getRemotepath()+rfilename+'\\'+fileName+suffix;
+ //String remote3=ftp.getRemotepath()+'\\'+machineModel.getMachine().getMachineName()+'\\'+machineModel.getOperation().getOperationName()+'\\'+machineModel.getProduct().getProductName();
+                rfilename=ftp.getRemotepath()+'\\'+machineModel.getMachine().getMachineName()+'\\'+machineModel.getOperation().getOperationName()+'\\'+machineModel.getProduct().getProductName()+'\\'+fileName+suffix;
                 machineModel.setMachineModelPath(rfilename);
 //                machineModel.setMachineModelName(modelName);//重命名新的模板名称
                 baseMapper.updateById(machineModel);//更新machineModel后的上传路径
                 //加日志
-
+                for (String s1:list) {
+                    FileUtil.DeleteFolder(s1);//删除本地暂存的文件
+                }
             }else{//上传失败
                 return ResponseEty.failure(message);
             }
         }
-        for (String s1:list) {
-            FileUtil.DeleteFolder(s1);//删除本地暂存的文件
+        if (isSuccess){
+            return ResponseEty.success(message);
+        }{
+            return ResponseEty.failure(message);
         }
 
-        return ResponseEty.success("文件上传成功");
+    }
+
+    @Override
+    public ResponseEty uploadTemplate1(MultipartFile[] file, Integer machineModelId,HttpServletRequest request, Integer tokenId) {
+        List<String> list = new ArrayList<String>();
+        String[] suffixs=new String[file.length];
+        if (file.length>1){
+            return ResponseEty.failure("只能上传一个文件");
+        }
+        MachineModel machineModel=baseMapper.selectById(machineModelId);
+        machineModel.setMachine(machineMapper.selectById(machineModel.getMachineId()));
+        machineModel.setOperation(operationMapper.selectById(machineModel.getOperationId()));
+        machineModel.setProduct(productsMapper.selectById(machineModel.getProductId()));
+        String rfilename=machineModel.getMachineModelName();
+        machineModel.setMachineModelNum(machineModel.getMachineModelNum()+1);//版本号+1
+        String fileName=machineModel.getMachineModelNum()+rfilename;
+        machineModel.setFileName(fileName);
+        if ( file.length > 0) {
+            for (int i = 0; i < file.length; i++) {
+                MultipartFile files = file[i];
+
+// 保存文件
+                list = FileUtil.saveFile(request, files, list,ftp,fileName);
+                suffixs[i]=FileUtil.getSuffix(files);
+            }
+        }
+        System.out.println(file.length);
+
+        String s; // 本地文件的全地址
+        String suffix;//文件的后缀
+        Boolean isSuccess = true;
+        String message="";
+        for (int j=0;j<list.size();j++){
+            FTPClient ftpClient = FTPUtil.connectFtpServer(ftp.getIpAddr(), ftp.getPort(), ftp.getUserName(), ftp.getPwd(), ftp.getEncoding());
+//            FTPUtil ftpUtil=new FTPUtil();
+            s=list.get(j);
+            suffix=suffixs[j];
+
+//            String modelName=machineModel.getMachineModelName();
+
+
+            Map map = FTPUtil.uploadMachineModelFiles(ftpClient, new File(s),ftp,machineModel);
+
+            isSuccess= (Boolean) map.get("success");
+            message= (String) map.get("message");
+            if (isSuccess){//上传成功
+                FTPUtil.closeFTPConnect(ftpClient);
+                //String remote3=ftp.getRemotepath()+'\\'+machineModel.getMachine().getMachineName()+'\\'+machineModel.getOperation().getOperationName()+'\\'+machineModel.getProduct().getProductName();
+                rfilename=ftp.getRemotepath()+'\\'+machineModel.getMachine().getMachineName()+'\\'+machineModel.getOperation().getOperationName()+'\\'+machineModel.getProduct().getProductName()+'\\'+fileName+suffix;
+                machineModel.setMachineModelPath(rfilename);
+//                machineModel.setMachineModelName(modelName);//重命名新的模板名称
+                baseMapper.updateById(machineModel);//更新machineModel后的上传路径
+                //加日志
+                for (String s1:list) {
+                    FileUtil.DeleteFolder(s1);//删除本地暂存的文件
+                }
+                return ResponseEty.success("上传成功");
+            }else{//上传失败
+                return ResponseEty.failure(message);
+            }
+        }
+        return ResponseEty.failure("上传失败");
+  /*      if (isSuccess){
+            return ResponseEty.success("上传成功");
+        }{
+            return ResponseEty.failure("上传失败");
+        }*/
     }
 
     @Override
